allowed custom vision data loaders with additional kwargs. fixed po synchronizer
